♻️ refactor: Login Modal 컴포넌트 분리, Card 컴포넌트화, Header 컴포넌트 추가작업, Swiper 컴포넌트화 Login 불필요한 상태 제거 및 전역 상태 타입추가
