--- src/components/Logo/Logo.jsx
@@ -1,9 +1,0 @@
-import * as Styled from "./Logo.styled";
-
-export default function Logo() {
-  return (
-    <Styled.Logo>
-      Min<span>Flix</span>
-    </Styled.Logo>
-  );
-}
