+++ src/components/Logo/index.jsx
@@ -1,0 +1,9 @@
+import { LogoWrapper } from './style'
+
+export default function Logo() {
+  return (
+    <LogoWrapper>
+      Min<span>Flix</span>
+    </LogoWrapper>
+  )
+}
