+++ src/components/Swiper/SwiperSection/style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+export const SectionWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-top: 5rem;
+  max-width: 1366px;
+
+  .title_box {
+    margin: 0 0 24px 0;
+  }
+`
