+++ src/components/Swiper/SwiperVideos/index.jsx
@@ -1,22 +1,20 @@
-import React, { useRef, useState } from "react";
+import React from 'react'
 // Import Swiper React components
-import { Swiper, SwiperSlide } from "swiper/react";
+import { Swiper, SwiperSlide } from 'swiper/react'
 
 // // Import Swiper styles
-import "swiper/css";
-import "swiper/css/pagination";
-import "swiper/css/navigation";
-
-import "../App.scss";
+import 'swiper/css'
+import 'swiper/css/pagination'
+import 'swiper/css/navigation'
 
 // import required modules
-import { Pagination, Navigation } from "swiper/modules";
+import { Pagination, Navigation } from 'swiper/modules'
 
 export default function SwiperVideos({ videos }) {
   return (
     <>
       <Swiper
-        slidesPerView={"auto"}
+        slidesPerView={'auto'}
         spaceBetween={30}
         pagination={{
           clickable: true,
@@ -39,5 +37,5 @@ export default function SwiperVideos({ videos }) {
           ))}
       </Swiper>
     </>
-  );
+  )
 }
