+++ src/components/UserProfile/style.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+export const ProfileWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  width: 100%;
+  cursor: pointer;
+
+  img {
+    width: 36px;
+    height: 36px;
+    border-radius: 50%;
+  }
+`
