+++ src/main.jsx
@@ -4,7 +4,7 @@ import './index.css'
 import { BrowserRouter } from 'react-router-dom'
 import { Provider } from 'react-redux'
 import { store } from './store/store.jsx'
-import Toast from './components/Toast.jsx'
+import Toast from './components/Toast/index.jsx'
 
 createRoot(document.getElementById('root')).render(
   <Provider store={store}>
