+++ src/pages/Search.jsx
@@ -1,60 +1,60 @@
-import { useEffect, useState } from "react";
-import { clientMovie } from "../client/clientMovie";
-import Card from "../components/Card";
+import { useEffect, useState } from 'react'
+import { clientMovie } from '../client/clientMovie'
+import Card from '../components/Card/Card'
 
-import "../styles/search.scss";
-import "../styles/home.scss";
+import '../styles/search.scss'
+import '../styles/home.scss'
 
 function Search() {
-  const [value, setValue] = useState("");
-  const [data, setData] = useState([]);
-  const [category, setCategory] = useState("movie");
+  const [value, setValue] = useState('')
+  const [data, setData] = useState([])
+  const [category, setCategory] = useState('movie')
 
-  window.scrollTo(0, 0);
+  window.scrollTo(0, 0)
 
   const onChange = (event) => {
-    const { value } = event.target;
-    setValue(value);
-  };
+    const { value } = event.target
+    setValue(value)
+  }
 
   useEffect(() => {
     const id = setTimeout(() => {
       switch (category) {
-        case "movie":
+        case 'movie':
           clientMovie
             .get(`/search/movie?query=${value}`)
             .then((res) => {
-              console.log(res.data.results);
-              setData(res.data.results);
+              console.log(res.data.results)
+              setData(res.data.results)
             })
-            .catch((err) => console.err(err));
-          break;
+            .catch((err) => console.err(err))
+          break
 
-        case "TV":
+        case 'TV':
           clientMovie
             .get(`/search/tv?query=${value}`)
             .then((res) => {
-              console.log(res.data.results);
-              setData(res.data.results);
+              console.log(res.data.results)
+              setData(res.data.results)
             })
-            .catch((err) => console.err(err));
-          break;
+            .catch((err) => console.err(err))
+          break
 
         // 배우, card 컴포넌트 프롭스 타입 추가 및 추가 작성 코드 필요
-        case "actor":
+        case 'actor':
           clientMovie
             .get(`/search/person?query=${value}`)
             .then((res) => {
-              console.log(res.data.results);
-              setData(res.data.results);
+              console.log(res.data.results)
+              setData(res.data.results)
             })
-            .catch((err) => console.err(err));
-          break;
+            .catch((err) => console.err(err))
+          break
       }
-    }, 1000);
+    }, 1000)
 
-    return () => clearTimeout(id);
-  }, [value]);
+    return () => clearTimeout(id)
+  }, [value])
 
   return (
     <>
@@ -62,21 +62,21 @@ function Search() {
         <div className="search_inner">
           <div className="search_category">
             <button
-              className={category === "movie" ? "btn_color" : null}
-              onClick={() => setCategory("movie")}
+              className={category === 'movie' ? 'btn_color' : null}
+              onClick={() => setCategory('movie')}
             >
               영화
             </button>
             <button
-              className={category === "TV" ? "btn_color" : null}
-              onClick={() => setCategory("TV")}
+              className={category === 'TV' ? 'btn_color' : null}
+              onClick={() => setCategory('TV')}
             >
               TV
             </button>
             {/* 배우, card 컴포넌트 프롭스 타입 추가 및 추가 작성 코드 필요 */}
             <button
-              className={category === "actor" ? "btn_color" : null}
-              onClick={() => setCategory("actor")}
+              className={category === 'actor' ? 'btn_color' : null}
+              onClick={() => setCategory('actor')}
             >
               배우
             </button>
@@ -95,7 +95,7 @@ function Search() {
         </div>
       </div>
     </>
-  );
+  )
 }
 
-export default Search;
+export default Search
