+++ src/styles/home.scss
@@ -23,72 +23,6 @@
 }
 
 // Card component START
-.card_container {
-  display: flex;
-  flex-direction: column;
-  margin-top: 5rem;
-  max-width: 1366px;
-}
-
-.title_box {
-  margin: 0 0 24px 0;
-}
-
-.card {
-  position: relative;
-  cursor: pointer;
-
-  &:hover {
-    .card_backdrop {
-      opacity: 1;
-      transition: 0.3s;
-    }
-
-    .card_more {
-      opacity: 1;
-      transition: 0.3s;
-    }
-  }
-
-  .card_backdrop {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    bottom: 0;
-    background: linear-gradient(0deg, black, transparent);
-    opacity: 0;
-    transition: 0.3s;
-  }
-
-  .card_more {
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    bottom: 0;
-    width: 100%;
-    height: max-content;
-    padding: 12px;
-    opacity: 0;
-    transition: 0.3s;
-    color: white;
-
-    .card_more_average {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      width: 48px;
-      height: 48px;
-      border-radius: 50%;
-      border: 2px solid rgb(32, 190, 98);
-
-      span {
-        font-size: 12px;
-        text-align: center;
-      }
-    }
-  }
-}
 
 .next_button {
   position: absolute;
@@ -102,64 +36,3 @@
   }
 }
 // Card component END
-
-.detail_sec_info2_container {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  padding-left: 10rem;
-
-  display: flex;
-  align-items: center;
-
-  .detail_sec_info2 {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 36px;
-    width: 60%;
-
-    .detail_sec_title2 {
-      font-size: 64px;
-      font-weight: 900;
-    }
-
-    .detail_sec_position2 {
-      display: flex;
-      align-items: center;
-      gap: 12px;
-
-      .detail_sec_position_list2 {
-        padding: 8px 12px;
-        background-color: red;
-        color: #fff;
-        border-radius: 12px;
-      }
-    }
-
-    .detail_sec_des2 {
-      line-height: 2;
-    }
-
-    button {
-      padding: 12px;
-      border-radius: 6px;
-
-      img {
-        fill: #fff;
-        stroke: #fff;
-      }
-    }
-  }
-}
-
-@media (min-width: 1200px) {
-  // .detail_sec_info2 {
-  // }
-
-  .detail_sec_title2 {
-    font-size: 36px;
-  }
-}
